move region fixture definition one level down to tests/aws
