+++ tests/conftest.py
@@ -5,10 +5,6 @@ import pytest
 from _pytest.config import PytestPluginManager
 from _pytest.config.argparsing import Parser
 
-from localstack.constants import TEST_AWS_REGION_NAME
-from localstack.testing.aws.util import is_aws_cloud
-from localstack.utils.bootstrap import is_api_enabled
-
 if TYPE_CHECKING:
     pass
 
@@ -169,11 +165,3 @@ def secondary_aws_client(aws_client_factory):
     from localstack.testing.aws.util import secondary_testing_aws_client
 
     return secondary_testing_aws_client(aws_client_factory)
-
-
-@pytest.fixture(name="region", scope="session")
-def fixture_region(aws_client):
-    if is_aws_cloud() or is_api_enabled("sts"):
-        return aws_client.sts.meta.region_name
-    else:
-        return TEST_AWS_REGION_NAME
